Rename service (and package) to singular form
